Move node mapping logic to a hook and restructure into sub functions
